+++ home-manager/default.nix
@@ -1,5 +1,6 @@
 {
   imports = [
     ./modules/git.nix
+    ./modules/zsh.nix
   ];
 }